Refactor Banner components
