--- src/components/atoms/Banner/Wrapper.js
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export default styled.div`
-  width: 100%;
-  height: 18.4rem;
-`;
